+++ src/torchmetrics/regression/log_mse.py
@@ -27,6 +27,15 @@ class MeanSquaredLogError(Metric):
 
     Where :math:`y` is a tensor of target values, and :math:`\hat{y}` is a tensor of predictions.
 
+    As input to ``forward`` and ``update`` the metric accepts the following input:
+
+    - ``preds`` (:class:`~torch.Tensor`): Predictions from model
+    - ``target`` (:class:`~torch.Tensor`): Ground truth values
+
+    As output of ``forward`` and ``compute`` the metric returns the following output:
+
+    - ``mean_squared_log_error`` (:class:`~torch.Tensor`): A tensor with the mean squared log error
+
     Args:
         kwargs: Additional keyword arguments, see :ref:`Metric kwargs` for more info.
 
@@ -57,12 +66,7 @@ class MeanSquaredLogError(Metric):
         self.add_state("total", default=tensor(0), dist_reduce_fx="sum")
 
     def update(self, preds: Tensor, target: Tensor) -> None:  # type: ignore
-        """Update state with predictions and targets.
-
-        Args:
-            preds: Predictions from model
-            target: Ground truth values
-        """
+        """Update state with predictions and targets."""
         sum_squared_log_error, n_obs = _mean_squared_log_error_update(preds, target)
 
         self.sum_squared_log_error += sum_squared_log_error